fix(win32): Throttle Vulkan rendering to vsync
The Vulkan swapchain uses MAILBOX present mode, which is tear-free but
does not block, so the dedicated render thread spun unthrottled (~2000
fps), rendering frames the compositor just discards. Pace it to the DWM
vsync via DwmFlush, keeping MAILBOX's low latency.

Extract the DwmFlush + degraded-timer fallback (previously inline in the
software renderer) into a shared Win32VSyncPacer used by both the
software and Vulkan renderers.
