+++ src/Uno.UI.Runtime.Skia.Win32/Rendering/Win32WindowWrapper.Rendering.Vulkan.cs
@@ -24,6 +24,7 @@ internal partial class Win32WindowWrapper
 	private class VulkanRenderer : IRenderer
 	{
 		private readonly HWND _hwnd;
+		private readonly Win32VSyncPacer _vsyncPacer = new(FeatureConfiguration.CompositionTarget.FrameRate);
 		private VulkanContext _vulkanContext;
 		private IDisposable? _deviceLock;
 		private bool _skipPresent;
@@ -111,6 +112,8 @@ internal partial class Win32WindowWrapper
 			if (_vulkanContext.CachedSkSurface == null || _vulkanContext.GrContext == null)
 				return;
 
+			_vsyncPacer.OnFrameStart();
+
 			// Flush Skia commands to the Vulkan intermediate image
 			_vulkanContext.CachedSkSurface.Canvas.Flush();
 			_vulkanContext.GrContext.Flush();
@@ -119,6 +122,11 @@ internal partial class Win32WindowWrapper
 			{
 				// Blit intermediate image to swapchain and present
 				_vulkanContext.BlitAndPresent();
+
+				// MAILBOX present returns without blocking, so pace the render thread to the
+				// compositor's vsync here; otherwise it spins at thousands of fps rendering frames
+				// the presentation engine discards. MAILBOX keeps latency low and stays tear-free.
+				_vsyncPacer.WaitForVSync();
 			}
 			_skipPresent = false;
 		}
@@ -150,10 +158,12 @@ internal partial class Win32WindowWrapper
 		{
 			_deviceLock?.Dispose();
 			_deviceLock = null;
+			_vsyncPacer.Dispose();
 			_vulkanContext.Dispose();
 		}
 
-		// Vulkan presents through its own VSync-aware swapchain, so no retargeting is needed here.
-		public void UpdateRefreshRate(double fps) { }
+		// Retargets the vsync pacer's degraded timer fallback (MAILBOX itself ignores this);
+		// mirrors the software renderer for the SetFrameRateAsScreenRefreshRate path.
+		public void UpdateRefreshRate(double fps) => _vsyncPacer.UpdateTargetFps(fps);
 	}
 }
